Fixes unicode/decode error happening in Python 2.7 use cases, when working with a string representation of a `Vote` model instance.
Context:
The `__str__` method associated to the `Vote` model includes the output of either the `__str__` or `__unicode__` that belong to the model upon which the voting acts on (referenced in the `content_object` field).

The `__str__` method associated to the `Vote` model is used internally by the `python_2_unicode_compatible` decorator.

Issue description:
There was an issue dealing with unicode/decode errors related to the potential mixing of pure-ascii strings and unicode strings in the `__str__` method of the `Vote` model. The potential unicode string would be a string representation of the instance referenced in the `content_object` field of the `Vote` model.

Fix:
Force a unicode string in Python 2.7 to be returned from the `__str__` method on the `Vote` model. This is accomplished via the `text_type` helper method included in the `six` package.